Split existing functionality into submodules
